Drop support for collectd < 5.7
With llvm apt repo not distributing clang 3.9+ on ubuntu 14.04 or 16.04,
regenerating the bindings with bindgen became a lot harder. I attempted
to see if I could generate all of them from checking out collectd tags
or copying includes from the older distros to a newer one, but both of
these methods failed.

I'm not too worried about dropping support as the rhel 7 epel comes with
collectd 5.8. If someone is effected by this and wants to see continued
maintenance of collectd 5.4 and collectd 5.5 API versions, consider
contributing or supporting me.