Guard slice with null and length checks
Tests started to fail with

> panic hook error: plugin panicked: (library/core/src/panicking.rs:
156): unsafe precondition(s) violated: slice::from_raw_parts requires
the pointer to be aligned and non-null, and the total size of the slice
not to exceed `isize::MAX`

It's not clear to me what changed to cause this (a bit of trial and
error showed the root cause being a `null` "children" or "values" in an
`oconfig_item_t`).

This commit fixes the issue by guarding slice creation with null and
length checks.